coverage: Move unused-function synthesis into its own submodule
